+++ changelog.md
@@ -12,7 +12,7 @@
 [//]: # "Additions:"
 
 - Added `newStringUninit` to system, which creates a new string of length `len` like `newString` but with uninitialized content.
-- Added `hasDefault` to `std/typetraits` to check if a type has a valid default value.
+- Added `hasDefaultValue` to `std/typetraits` to check if a type has a valid default value.
 
 [//]: # "Deprecations:"